Move target integration tests under integration
