+++ tests/integration/local_h5/__init__.py
@@ -1,0 +1,1 @@
+"""Local H5 integration tests."""
