--- tests/optimized/local_h5/__init__.py
@@ -1,1 +1,0 @@
-"""Optimized local H5 integration tests."""
